(#46) Ensure run tests from separate package
We want our tests to be higher level as possible so it
mimics a library user as much as possible. Also, we want
to make it easy and obvious the fact that we're not
including tests in the distribution.

Ensure `python -m build` does not include test codes
in the distribution:

```bash
source .venv/bin/activate \
&& rm -rf dist/ \
&& pip uninstall aiotaskq tests -y \
&& python -m build \
&& pip install --quiet dist/aiotaskq-0.0.11-py3-none-any.whl \
&& [ $(ls -l .venv/lib/python3.10/site-packages/aiotaskq | grep tests | grep '^d.*') ] \
&& echo "Fail" \
|| echo "Pass"
```
