+++ .pylintrc.tests
