+++ .vscode/settings.json
@@ -9,5 +9,8 @@
         "**/.mypy_cache": true
     },
     "python.testing.pytestEnabled": true,
-    "python.testing.unittestEnabled": false
+    "python.testing.unittestEnabled": false,
+    "python.analysis.extraPaths": [
+        "src/tests/"
+    ]
 }
