+++ lint.sh
@@ -3,10 +3,10 @@ pip install -e .[dev]
 
 if [ -z $1 ];
 then
-    pylint -v --rcfile ./.pylintrc --ignore-paths=src/aiotaskq/tests/* src/aiotaskq
+    pylint -v --rcfile ./.pylintrc src/aiotaskq
     failed_2=$?
 
-    pylint -v --rcfile ./.pylintrc.test src/aiotaskq/tests/
+    pylint -v --rcfile ./.pylintrc.tests src/tests/
     failed_1=$?
 
     [ $failed_1 -eq 0 ] && [ $failed_2 -eq 0 ] && failed=0 || failed=1
